Cleanup mock services
